@@ -1,6 +1,6 @@
 import React from 'react';
 import RunDetailsModal from './index';
-import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
+import Adapter from '@cfaester/enzyme-adapter-react-18';
 import { configure } from 'enzyme';
 import { waitFor } from '@testing-library/react';
 import { ButtonTimeoutContext } from '../../../utils/button-timeout-context';
@@ -62,7 +62,9 @@ describe('RunDetailsModal', () => {
 
     onClick.mockImplementation((visible) => [visible, setVisible]);
 
-    closeButton.simulate('click');
+    waitFor(() => {
+      closeButton.simulate('click');
+    });
 
     expect(
       wrapper.find(
