@@ -1,7 +1,7 @@
 import React from 'react';
 import configureMockStore from 'redux-mock-store';
 import RunExportModal from './index';
-import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
+import Adapter from '@cfaester/enzyme-adapter-react-18';
 import { configure } from 'enzyme';
 import { render, screen } from '@testing-library/react';
 import { ButtonTimeoutContext } from '../../../utils/button-timeout-context';
