@@ -1,7 +1,7 @@
 import React from 'react';
 import configureMockStore from 'redux-mock-store';
 import RunsListCard from '.';
-import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
+import Adapter from '@cfaester/enzyme-adapter-react-18';
 import { configure, mount } from 'enzyme';
 import { HoverStateContext } from '../utils/hover-state-context';
 import { setup } from '../../../utils/state.mock';
