@@ -1,6 +1,6 @@
 import React from 'react';
 import Switch from './switch';
-import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
+import Adapter from '@cfaester/enzyme-adapter-react-18';
 import { configure, mount, shallow } from 'enzyme';
 
 configure({ adapter: new Adapter() });
